[InlineCost] Avoid inlining decisions that would cause miscompilation.
This patch adds a new method to TargetTransformInfo, which lets
targets decide whether a certain inline decision would surely lead to
miscompilation. If so, no inlining is performed even if the function
has always_inline attributes.

On X86, this function is implemented by calling into
areInlineCompatible, as that code is fairly precise on what would *not*
be compatible with inlining. Unfortunately other targets have
historically been not as precise, leading to some issues with a similar
PR in the past (https://reviews.llvm.org/D150396) that caused internal
compiler errors in some cases and ultimately lead to the PR being
reverted, so for now this check is disabled on non-x86 targets (i.e.
inliningWouldCauseMiscompilation always returns false).